Analyzing the choice of career for the Alumni of Engineering College using SQL
In this project, students are expected to analyze the career choices of alumni from an institution based on given data.

Once a student is passed out from an Institute or College, he/she is known as Alumni of the Institute. Alumni career growth plays important role in Institute’s ranking and other networking activities. In this project, the career choices of alumni of two Universities will be analyzed with respect to their passing year as well as the course they completed.